+++ jeap-spring-boot-audit-starter-transactional-outbox/src/test/java/ch/admin/bit/jeap/audit/transactional/outbox/TestApp.java
@@ -6,7 +6,13 @@ import ch.admin.bit.jeap.messaging.annotations.JeapMessageProducerContract;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@JeapMessageProducerContract(value = CreateAuditRecordCommand.TypeRef.class, topic = "mysystem-audit")
-@JeapMessageConsumerContract(value = CreateAuditRecordCommand.TypeRef.class, topic = "mysystem-audit")
+@JeapMessageProducerContract(value = CreateAuditRecordCommand.TypeRef.class, topic = {
+        "mysystem-audit",
+        "mysystem-audit-secondary"
+})
+@JeapMessageConsumerContract(value = CreateAuditRecordCommand.TypeRef.class, topic = {
+        "mysystem-audit",
+        "mysystem-audit-secondary"
+})
 class TestApp {
 }